requirements.txt
tensorflow==2.3.0
requests==2.24.0
flask==1.1.2
pyaml==20.4.0
python-dotenv==0.14.0